+++ packages/connectkit/src/components/Common/Tooltip/index.tsx
@@ -7,7 +7,7 @@ import { TooltipProps, TooltipSizeProps } from "./types";
 
 import { AnimatePresence, motion } from "framer-motion";
 import { ResetContainer } from "../../../styles";
-import { useThemeContext } from "../../DaimoPayThemeProvider/DaimoPayThemeProvider";
+import { useThemeContext } from "../../RozoPayThemeProvider/RozoPayThemeProvider";
 import Portal from "../Portal";
 
 const Tooltip: React.FC<TooltipProps> = ({
