+++ packages/connectkit/src/components/Pages/MobileConnectors/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { ROUTES } from "../../../constants/routes";
-import { useRozoPay } from "../../../hooks/useDaimoPay";
+import { useRozoPay } from "../../../hooks/useRozoPay";
 import { usePayContext } from "../../../hooks/usePayContext";
 import {
   WalletConfigProps,
