+++ packages/connectkit/src/components/RozoPayModal/index.tsx
@@ -5,14 +5,14 @@ import { useAccount } from "wagmi";
 import { ROUTES } from "../../constants/routes";
 import { getAppName } from "../../defaultConfig";
 import { useChainIsSupported } from "../../hooks/useChainIsSupported";
-import { useRozoPay } from "../../hooks/useDaimoPay";
+import { useRozoPay } from "../../hooks/useRozoPay";
 import useIsMobile from "../../hooks/useIsMobile";
 import { usePayContext } from "../../hooks/usePayContext";
 import { useStellar } from "../../provider/StellarContextProvider";
 import { CustomTheme, Languages, Mode, Theme } from "../../types";
 import { IntercomInitializer } from "../Common/Intercom";
 import Modal from "../Common/Modal";
-import { RozoPayThemeProvider } from "../DaimoPayThemeProvider/DaimoPayThemeProvider";
+import { RozoPayThemeProvider } from "../RozoPayThemeProvider/RozoPayThemeProvider";
 import About from "../Pages/About";
 import Confirmation from "../Pages/Confirmation";
 import Connectors from "../Pages/Connectors";
