+++ packages/connectkit/src/world-mini-app/WorldPayButton.tsx
@@ -22,9 +22,9 @@ import { Address, Hex } from "viem";
 import ThemedButton, {
   ThemeContainer,
 } from "../components/Common/ThemedButton";
-import { RozoPayButtonInner } from "../components/DaimoPayButton";
+import { RozoPayButtonInner } from "../components/RozoPayButton";
 import { ROUTES } from "../constants/routes";
-import { useRozoPay } from "../hooks/useDaimoPay";
+import { useRozoPay } from "../hooks/useRozoPay";
 import { usePayContext } from "../hooks/usePayContext";
 import { ResetContainer } from "../styles";
 import { CustomTheme, Mode, Theme } from "../types";
